requirements.txt
pytorch
# numpy
einops
chess